Canviades Act15

+++ CODE INSITUT CFGS/Programacio/Java/UD1/Ud1Act15MarcMas/src/main/java/org/yourcompany/yourproject/Ud1Act15MarcMas.java
@@ -16,12 +16,12 @@ public class Ud1Act15MarcMas {
         System.out.println("Calculadora de potes capturades");
         Scanner s = new Scanner(System.in); 
         System.out.println("Indiqui el nombre de formigues capturades ");
-        Short nFormigues = s.nextShort(); //agaf nombre per consola
+        Integer nFormigues = s.nextInt(); //agaf nombre per consola
         System.out.println("Indiqui el nombre de aranyes capturades");
-        Short nAranyes = s.nextShort(); //agaf nombre per consola
+        Integer nAranyes = s.nextInt(); //agaf nombre per consola
         System.out.println("Indiqui el nombre de centpeus capturades");
-        Short nCentpeus = s.nextShort(); //agaf nombre per consola
-        Float sumaPotes = (float) nFormigues * 6 + (float) nAranyes * 8 + (float) nCentpeus * 100; //faig la suma de totes les potes
+        Integer nCentpeus = s.nextInt(); //agaf nombre per consola
+        Integer sumaPotes = (nFormigues * 6) + (nAranyes * 8) + (nCentpeus * 100); //faig la suma de totes les potes
         System.out.println("La suma de les potes es " + sumaPotes); // imprimesc per pantalla
     }
 }
